collect the logic. also, focus to the first cell if change to the select lyric.
